Code Review (https://github.com/code4romania/monitorizare-vot/pull/319#discussion_r507587684): moved Interfaces and partial classes to respect https://documentation.help/StyleCop/SA1402.html
