+++ src/api/VoteMonitor.Api.DataExport/FileGenerator/CsvUtility.cs
@@ -1,0 +1,7 @@
+﻿namespace VoteMonitor.Api.DataExport.FileGenerator
+{
+    public class CsvUtility
+    {
+        public const string CSV_MEDIA_TYPE = "text/csv"; // ref: https://stackoverflow.com/questions/7076042/what-mime-type-should-i-use-for-csv > https://tools.ietf.org/html/rfc7111
+    }
+}
