+++ src/api/VoteMonitor.Api.DataExport/FileGenerator/ICsvGenerator.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace VoteMonitor.Api.DataExport.FileGenerator
+{
+    public interface ICsvGenerator
+    {
+        byte[] Export<T>(IEnumerable<T> exportData, string fileName);
+    }
+}
